Add mock backend
There's still a problem where the backend's state is not isolatable, causing tests to fail. For dev mode, however, it works just fine. My idea is to fix the isolation issue by letting the test driver pass the state to the server. If that doesn't work, I may have to bring back mock-socket as a proxy.